# packages/schemas/openapi.v1.yaml
openapi: 3.1.0
info:
  title: MARID Control Plane API
  version: 0.3.0
  description: Tenant-scoped control plane. Execution records do not execute tools.
servers:
  - url: http://localhost:5056
paths:
...
  /v1/deployment/domains:
    get:
      summary: List verified-domain and certificate status for the token tenant
      description: Requires AUDITOR or INCIDENT_MANAGER role. Verification tokens are not returned.
      security: [{ bearerAuth: [] }]
      x-required-scope: deployment.domains.read
      responses:
        '200': { description: Tenant domain status list }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant role }
  /v1/events:
    post:
      summary: Ingest one normalized event; source and sourceEventId are tenant-idempotent
      description: Requires an enabled tenant principal with the EVENT_INGESTOR role. Reusing a source event ID with changed content returns 409 and is audited.
      security: [{ bearerAuth: [] }]
      x-required-scope: events.ingest
      parameters:
        - in: header
          name: X-Correlation-ID
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./security-event.v1.schema.json
      responses:
        '201': { description: Event created }
        '200': { description: Duplicate event already present }
        '400': { description: Invalid envelope }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant context }
        '409': { description: Source event ID already exists with different content }
  /v1/incidents:
    get:
      summary: List the most recent 100 incidents for the token tenant
      description: Requires ANALYST, INCIDENT_MANAGER, RESPONDER, APPROVER, or AUDITOR role.
      security: [{ bearerAuth: [] }]
      x-required-scope: incidents.read
      responses:
        '200': { description: Tenant incident list }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant context }
    post:
      summary: Create a tenant incident
      description: Requires INCIDENT_MANAGER role.
      security: [{ bearerAuth: [] }]
      x-required-scope: incidents.write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [title, severity]
              properties:
                title: { type: string, minLength: 1, maxLength: 200 }
                severity: { type: string, enum: [Low, Medium, High, Critical] }
      responses:
        '201': { description: Incident created }
        '400': { description: Invalid incident }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant context }
  /v1/response/proposals:
    get:
      summary: List the latest 100 tenant response proposals and review decisions
      description: Requires RESPONDER, APPROVER, or AUDITOR role.
      security: [{ bearerAuth: [] }]
      x-required-scope: response.read
      responses:
        '200': { description: Tenant proposal list }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant context }
    post:
      summary: Evaluate policy and record a response proposal; does not execute it
      description: Requires RESPONDER role.
      security: [{ bearerAuth: [] }]
      x-required-scope: response.propose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./response-proposal.v1.schema.json
      responses:
        '201': { description: Proposal recorded }
        '200': { description: Policy denied, deferred, or requested more evidence }
        '400': { description: Invalid proposal }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope or tenant context }
  /v1/response/proposals/{proposalId}/decisions:
    post:
      summary: Record one human review decision for an exact proposal; does not execute it
      security: [{ bearerAuth: [] }]
      x-required-scope: response.approve
      description: Requires actor_type=human, an APPROVER tenant role, and a reviewer distinct from the proposer. The approval binds the full action hash, including tenant, capability, resource, risk, parameters, and linked event IDs.
      parameters:
        - name: proposalId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APPROVE, REJECT, REQUEST_MORE_EVIDENCE]
      responses:
        '201': { description: Decision recorded }
        '400': { description: Invalid decision }
        '401': { description: Missing or invalid token }
        '403': { description: Missing scope, human claim, or tenant context }
        '404': { description: Proposal not found in this tenant }
        '409': { description: Expired, self reviewed, altered, or already decided proposal }
...
  /v1/operator/sessions:
    post:
      summary: Start an audited Nocturne engineer session with ticket and reason
      security: [{ bearerAuth: [] }]
      x-required-scope: operator.session
      responses:
        '201': { description: Session started }
        '403': { description: Requires signed engineer class and tenant role }
...
  /v1/incidents/{incidentId}/activity:
    get:
      summary: Read tenant incident history
      security: [{ bearerAuth: [] }]
      x-required-scope: incidents.read
      parameters:
        - { name: incidentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Incident activity }
        '404': { description: No incident in the token tenant }
    post:
      summary: Append an investigation, remediation, or status transition
      security: [{ bearerAuth: [] }]
      x-required-scope: incidents.write
      parameters:
        - { name: incidentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Activity recorded }
        '404': { description: No incident in the token tenant }
        '409': { description: Invalid status transition }
  /v1/incidents/{incidentId}/events/{eventId}:
    post:
      summary: Link tenant evidence event to tenant incident
      security: [{ bearerAuth: [] }]
      x-required-scope: incidents.write
      parameters:
        - { name: incidentId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Link created }
        '404': { description: Incident or event is not in the token tenant }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT